# pkg/api/openapi.yaml
openapi: 3.0.3
info:
  title: Orcinus API
  description: |
    REST API for orcinus — Compose-simple, Cluster-strong. Deploy docker-compose
    files and Kubernetes manifests, manage projects, scale, secrets, plugins and
    inspect the cluster. Served by `orcinus api`.

    All `/api/v1/*` routes require `Authorization: Bearer <token>` when the server
    was started with a token; `/healthz`, `/version`, `/openapi.json` and `/docs`
    are always open.
  version: "1.0"
servers:
  - url: /
security:
  - bearerAuth: []
tags:
  - name: system
  - name: deploy
  - name: projects
  - name: secrets
  - name: plugins
  - name: cluster
paths:
  /healthz:
    get:
      tags: [system]
      summary: Liveness probe
      security: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties: { status: { type: string, example: ok } }
  /version:
    get:
      tags: [system]
      summary: Build & component versions
      security: []
      responses:
        "200":
          description: version info
          content:
            application/json:
              schema:
                type: object
                properties:
                  version: { type: string }
                  gitCommit: { type: string }
                  kompose: { type: string }
  /api/v1/convert:
    post:
      tags: [deploy]
      summary: Render input to Kubernetes manifests (no apply)
      description: Accepts a JSON DeployRequest, or a raw compose/manifest body with options as query params.
      parameters:
        - { name: project, in: query, schema: { type: string } }
        - { name: namespace, in: query, schema: { type: string } }
        - { name: mode, in: query, schema: { type: string, enum: [compose, manifest] } }
        - { name: replicas, in: query, schema: { type: integer } }
        - { name: pvcSize, in: query, schema: { type: string, example: 1Gi } }
        - { name: prune, in: query, schema: { type: boolean, default: true } }
        - { name: prunePVCs, in: query, schema: { type: boolean, default: false } }
        - { name: wait, in: query, schema: { type: boolean, default: false } }
        - { name: acmeEmail, in: query, schema: { type: string } }
        - { name: profile, in: query, description: repeatable, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DeployRequest" }
          application/yaml:
            schema: { type: string }
      responses:
        "200":
          description: rendered manifests
          content:
            application/json:
              schema:
                type: object
                properties:
                  objects: { type: integer }
                  manifests: { type: string, description: multi-document YAML }
        "400": { $ref: "#/components/responses/Error" }
  /api/v1/deploy:
    post:
      tags: [deploy]
      summary: Convert and apply input to the cluster
      description: |
        Detects compose vs manifest per document, converts, auto-installs any
        required plugins (cert-manager for x-orcinus-tls, argo-rollouts for
        x-orcinus-rollout), then server-side-applies with ownership labels.
      parameters:
        - { name: project, in: query, schema: { type: string } }
        - { name: namespace, in: query, schema: { type: string } }
        - { name: mode, in: query, schema: { type: string, enum: [compose, manifest] } }
        - { name: replicas, in: query, schema: { type: integer } }
        - { name: pvcSize, in: query, schema: { type: string, example: 1Gi } }
        - { name: prune, in: query, schema: { type: boolean, default: true } }
        - { name: prunePVCs, in: query, schema: { type: boolean, default: false } }
        - { name: wait, in: query, schema: { type: boolean, default: false } }
        - { name: acmeEmail, in: query, schema: { type: string } }
        - { name: profile, in: query, description: repeatable, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DeployRequest" }
          application/yaml:
            schema: { type: string }
      responses:
        "200":
          description: applied
          content:
            application/json:
              schema:
                type: object
                properties:
                  applied: { type: integer }
                  project: { type: string }
                  installed: { type: array, items: { type: string } }
        "400": { $ref: "#/components/responses/Error" }
  /api/v1/projects:
    get:
      tags: [projects]
      summary: List deployed projects
      parameters:
        - { name: namespace, in: query, schema: { type: string }, description: limit to one namespace }
      responses:
        "200":
          description: projects
          content:
            application/json:
              schema:
                type: object
                properties:
                  projects:
                    type: array
                    items: { $ref: "#/components/schemas/ProjectSummary" }
  /api/v1/projects/{project}:
    delete:
      tags: [projects]
      summary: Remove a project's resources
      parameters:
        - { name: project, in: path, required: true, schema: { type: string } }
        - { name: namespace, in: query, schema: { type: string } }
      responses:
        "200":
          description: removed count
          content:
            application/json:
              schema:
                type: object
                properties: { removed: { type: integer } }
  /api/v1/projects/{project}/pods:
    get:
      tags: [projects]
      summary: List a project's pods
      parameters:
        - { name: project, in: path, required: true, schema: { type: string } }
        - { name: namespace, in: query, schema: { type: string } }
      responses:
        "200":
          description: pods
          content:
            application/json:
              schema:
                type: object
                properties:
                  pods:
                    type: array
                    items: { $ref: "#/components/schemas/PodInfo" }
  /api/v1/projects/{project}/config:
    get:
      tags: [projects]
      summary: Read a project's live configuration, or export it
      description: |
        Reads back every orcinus-owned object of the project. `format=summary`
        gives a human-readable overview, `k8s` a re-appliable multi-document
        manifest stream, `orcinus` an orcinus.yml rebuilt from the deployed
        workloads. Secret values are always redacted on this endpoint.
      parameters:
        - { name: project, in: path, required: true, schema: { type: string } }
        - { name: namespace, in: query, schema: { type: string } }
        - name: format
          in: query
          schema: { type: string, enum: [summary, k8s, orcinus], default: summary }
      responses:
        "200":
          description: project configuration
          content:
            application/json:
              schema:
                type: object
                properties:
                  project: { type: string }
                  namespace: { type: string }
                  resources: { type: integer }
                  format: { type: string }
                  config: { type: string, description: the rendered configuration }
        "404": { $ref: "#/components/responses/Error" }
  /api/v1/projects/{project}/services/{service}/scale:
    post:
      tags: [projects]
      summary: Scale a service
      parameters:
        - { name: project, in: path, required: true, schema: { type: string } }
        - { name: service, in: path, required: true, schema: { type: string } }
        - { name: namespace, in: query, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties: { replicas: { type: integer } }
              required: [replicas]
      responses:
        "200":
          description: scaled
          content:
            application/json:
              schema:
                type: object
                properties:
                  scaled: { type: string }
                  replicas: { type: integer }
  /api/v1/projects/{project}/services/{service}/restart:
    post:
      tags: [projects]
      summary: Restart a service's pods (rolling; the spec is unchanged)
      parameters:
        - { name: project, in: path, required: true, schema: { type: string } }
        - { name: service, in: path, required: true, schema: { type: string } }
        - { name: namespace, in: query, schema: { type: string } }
      responses:
        "200":
          description: restarted
          content:
            application/json:
              schema:
                type: object
                properties: { restarted: { type: string, description: the kind that was restarted } }
  /api/v1/projects/{project}/services/{service}/rollback:
    post:
      tags: [projects]
      summary: Roll a service back to its previous revision
      parameters:
        - { name: project, in: path, required: true, schema: { type: string } }
        - { name: service, in: path, required: true, schema: { type: string } }
        - { name: namespace, in: query, schema: { type: string } }
      responses:
        "200":
          description: rolled back
          content:
            application/json:
              schema:
                type: object
                properties: { result: { type: string } }
  /api/v1/secrets:
    get:
      tags: [secrets]
      summary: List secrets
      parameters:
        - { name: namespace, in: query, schema: { type: string } }
      responses:
        "200":
          description: secrets
          content:
            application/json:
              schema:
                type: object
                properties:
                  secrets:
                    type: array
                    items: { $ref: "#/components/schemas/SecretInfo" }
    post:
      tags: [secrets]
      summary: Create or update an opaque secret
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SecretRequest" }
      responses:
        "201":
          description: created
          content:
            application/json:
              schema:
                type: object
                properties:
                  name: { type: string }
                  namespace: { type: string }
                  keys: { type: integer }
        "400": { $ref: "#/components/responses/Error" }
  /api/v1/secrets/docker-registry:
    post:
      tags: [secrets]
      summary: Create/update a private-registry pull secret (docker login)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegistrySecretRequest" }
      responses:
        "201":
          description: created
          content:
            application/json:
              schema:
                type: object
                properties:
                  name: { type: string }
                  namespace: { type: string }
                  server: { type: string }
        "400": { $ref: "#/components/responses/Error" }
  /api/v1/secrets/tls:
    post:
      tags: [secrets]
      summary: Create/update a TLS secret from an inline PEM cert + key
      description: >
        For a custom/BYO cert, referenced from compose with
        `x-orcinus-tls-secret`. The CLI equivalent reads files
        (`orcinus secret create-tls --cert --key`); over HTTP the PEM contents
        are sent inline.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TLSSecretRequest" }
      responses:
        "201":
          description: created
          content:
            application/json:
              schema:
                type: object
                properties:
                  name: { type: string }
                  namespace: { type: string }
                  type: { type: string, example: kubernetes.io/tls }
        "400": { $ref: "#/components/responses/Error" }
  /api/v1/secrets/{name}:
    get:
      tags: [secrets]
      summary: Show a secret's keys (values only with showValues)
      description: >
        Keys are listed and values withheld unless `showValues=true`, so a
        listing is safe to fetch without printing credentials. Mirrors
        `orcinus secret get`.
      parameters:
        - { name: name, in: path, required: true, schema: { type: string } }
        - { name: namespace, in: query, schema: { type: string } }
        - { name: showValues, in: query, schema: { type: boolean, default: false } }
      responses:
        "200":
          description: secret
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SecretDetail" }
        "404": { $ref: "#/components/responses/Error" }
    patch:
      tags: [secrets]
      summary: Set keys on a secret, keeping the ones not named
      description: >
        Merges the given keys, unlike `POST /api/v1/secrets` which replaces the
        data wholesale and drops keys it was not given. Creates the secret if it
        does not exist. Mirrors `orcinus secret set`. Running pods keep the old
        values until restarted, since env vars are injected at container start.
      parameters:
        - { name: name, in: path, required: true, schema: { type: string } }
        - { name: namespace, in: query, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SecretRequest" }
      responses:
        "200":
          description: updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  name: { type: string }
                  namespace: { type: string }
                  set: { type: integer, description: keys written by this call }
                  keys: { type: integer, description: total keys after the merge }
                  note: { type: string }
        "400": { $ref: "#/components/responses/Error" }
    delete:
      tags: [secrets]
      summary: Delete a secret
      parameters:
        - { name: name, in: path, required: true, schema: { type: string } }
        - { name: namespace, in: query, schema: { type: string } }
      responses:
        "204": { description: deleted }
  /api/v1/plugins:
    get:
      tags: [plugins]
      summary: List available plugins and install state
      responses:
        "200":
          description: plugins
          content:
            application/json:
              schema:
                type: object
                properties:
                  plugins:
                    type: array
                    items: { $ref: "#/components/schemas/PluginInfo" }
  /api/v1/plugins/{name}:
    post:
      tags: [plugins]
      summary: Install a plugin
      parameters:
        - { name: name, in: path, required: true, schema: { type: string } }
      requestBody:
        required: false
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PluginRequest" }
      responses:
        "200":
          description: installed
          content:
            application/json:
              schema:
                type: object
                properties: { installed: { type: string } }
        "404": { $ref: "#/components/responses/Error" }
    delete:
      tags: [plugins]
      summary: Remove a plugin
      parameters:
        - { name: name, in: path, required: true, schema: { type: string } }
      responses:
        "204": { description: removed }
        "404": { $ref: "#/components/responses/Error" }
  /api/v1/cluster:
    get:
      tags: [cluster]
      summary: Cluster status (from saved state)
      responses:
        "200":
          description: status
          content:
            application/json:
              schema:
                type: object
                properties:
                  name: { type: string }
                  running: { type: boolean }
                  server: { type: string }
                  runtime: { type: string }
                  kubeconfig: { type: string }
                  nodes: { type: string }
        "503": { $ref: "#/components/responses/Error" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses:
    Error:
      description: error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      properties: { error: { type: string } }
    DeployRequest:
      type: object
      required: [source]
      properties:
        source: { type: string, description: compose and/or manifest YAML (multi-doc) }
        project: { type: string, default: default }
        namespace: { type: string }
        mode: { type: string, enum: ["", compose, manifest] }
        replicas: { type: integer }
        pvcSize: { type: string, example: 1Gi }
        prune: { type: boolean, default: true }
        prunePVCs: { type: boolean, default: false, description: also delete the PersistentVolumeClaims of removed services (destroys their data) }
        wait: { type: boolean }
        acmeEmail: { type: string }
        profiles: { type: array, items: { type: string }, description: compose profiles to activate }
    ProjectSummary:
      type: object
      properties:
        Name: { type: string }
        Namespaces: { type: array, items: { type: string } }
        Workloads: { type: integer }
        Ready: { type: integer }
    PodInfo:
      type: object
      properties:
        Name: { type: string }
        Namespace: { type: string }
        Service: { type: string }
        Ready: { type: string }
        Status: { type: string }
        Restarts: { type: integer }
        Node: { type: string }
    SecretInfo:
      type: object
      description: >
        Field names are capitalised here for backwards compatibility; the rest
        of the API uses camelCase.
      properties:
        Name: { type: string }
        Type: { type: string }
        Keys: { type: integer, description: number of keys }
        KeyNames:
          type: array
          description: the key names themselves
          items: { type: string }
        ManagedBy: { type: boolean }
    SecretDetail:
      type: object
      properties:
        name: { type: string }
        namespace: { type: string }
        type: { type: string, example: Opaque }
        managedBy: { type: boolean }
        keys:
          type: array
          items: { type: string }
        values:
          type: object
          description: present only when showValues=true
          additionalProperties: { type: string }
        redacted:
          type: boolean
          description: present (true) when values were withheld
    TLSSecretRequest:
      type: object
      required: [name, cert, key]
      properties:
        name: { type: string }
        namespace: { type: string, default: default }
        cert: { type: string, description: PEM certificate chain }
        key: { type: string, description: PEM private key }
    SecretRequest:
      type: object
      required: [name, data]
      properties:
        name: { type: string }
        namespace: { type: string, default: default }
        data:
          type: object
          additionalProperties: { type: string }
    RegistrySecretRequest:
      type: object
      required: [name, server, username, password]
      properties:
        name: { type: string }
        namespace: { type: string, default: default }
        server: { type: string, example: registry.example.com }
        username: { type: string }
        password: { type: string }
        email: { type: string }
        insecure: { type: boolean, description: skip TLS verification during the login test }
        skipLoginCheck: { type: boolean, description: create without testing the login (400 if the login test fails) }
    PluginInfo:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
        version: { type: string }
        providers: { type: array, items: { type: string } }
        installed: { type: boolean }
    PluginRequest:
      type: object
      properties:
        email: { type: string }
        staging: { type: boolean }
        provider: { type: string }
        size: { type: string }
        replicas: { type: integer }
        nfsServer: { type: string }
        nfsPath: { type: string }